backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pydantic==2.8.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
resend==2.3.0
httpx==0.27.2
python-multipart==0.0.9
razorpay==1.4.2
